install_requirements.txt
# cl-runtime>=2.0.5
chromadb>=0.5.5
faiss-cpu>=1.8.0.post1
guidance>=0.1.15
langchain>=0.2.11
langchain-community>=0.2.10
langchain-fireworks>=0.1.5
langchain-openai>=0.1.19
openai>=1.37.1
python-dotenv>=1.0.1
